add assertIsOwner statement that was left off from delete sdb method when the code was ported from Karyon to Riposte

+++ src/main/java/com/nike/cerberus/service/SafeDepositBoxService.java
@@ -289,6 +289,8 @@ public class SafeDepositBoxService {
                     .build();
         }
 
+        assertIsOwner(groups, box.get());
+
         // 1. Remove permissions and metadata from database.
         iamRolePermissionService.deleteIamRolePermissions(id);
         userGroupPermissionService.deleteUserGroupPermissions(id);
